Create README.md
Smart Entrance QR System, worked as a team leader. involved in the design and coding of the module.

The main purpose of the system is to record student attendance via QR code. It works smartly, and the installation cost is very low because it operates on the Android platform.

This is a very basic application with working functionality to capture entry and exit attendance and store it in a Google Spreadsheet.

This application was built with the online Android app builder tool "Thunkable Classic."